Refactor: Extract SelectedAccountProvider to separate file
